+++ NEWS.md
@@ -6,7 +6,7 @@ emmeans 1.6.1.09xxxxx
     * Fixed navigation error in vignette index
     * Discouraging message added to `cld()` results. 
       Also am providing an `emm_list` method for `emm_list` objects.
-    * Added `mvcontrast()` function (#281)
+    * Added `mvcontrast()` function (#281) and assoc vignette material
     * Added `update.summary_emm()`